+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## [Unreleased]
 
+## [0.7] - 2018-03-22
+### Changed
+- entropy(): renamed argument 'dist' to 'plugin'
+
+## [0.6] - 2018-03-22
+### Changed
+- entropy(): renamed argument 'a' to 'alpha'
+
+## [0.5.5] - 2018-03-16
+### Changed
+- fix missing README.rst
+- handle input counts arrays > 1D (always flatten to 1D)
+- handle single-class case
+
 ## [0.5.3] - 2018-03-15
 ### Changed
 - require numpy >= 1.9